Wed, Nov 1, 2023, 12:19 AM +00:00

+++ CS50/week 2/scrabble/scrabble.c
@@ -24,7 +24,7 @@ int main(void)
 
     for (int i = 0; i < players; i++)
     {
-        scores[i] += calc_score(words[i]);
+        scores[i] += (int) calc_score(words[i]);
         printf("score: %i\n", scores[i]);
     }
